KAN-68
The most significant changes involve the `BaseDbModel` and `Savings` classes. The `BaseDbModel` class no longer inherits from the `BaseModel` class, which may affect its functionality and the way it interacts with other classes. Additionally, the `PrimaryKey` attribute for the `Id` property in the `BaseDbModel` class has been removed, which could impact how records are identified in the database. In the `Savings` class, the `Table` attribute has been removed, which may affect how the class is mapped to the database. Also, the data type of the `Amount` property in the `Savings` class has been changed from `float` to `double`, which could affect the precision of the values stored.

List of changes:

1. The `BaseDbModel` class no longer inherits from the `BaseModel` class. This could potentially change the behavior of the `BaseDbModel` class and its interaction with other classes. (See code change 1)

2. The `PrimaryKey` attribute for the `Id` property in the `BaseDbModel` class has been removed. This could affect how records are uniquely identified in the database. (See code change 2)

3. The `Table` attribute for the `Savings` class has been removed. This could potentially affect how the `Savings` class is mapped to the database. (See code change 3)

4. The data type of the `Amount` property in the `Savings` class has been changed from `float` to `double`. This could affect the precision of the values stored in the `Amount` property. (See code change 4)