KAN-67
The most significant changes include the addition of three new classes `Budget`, `BudgetOverview`, and `BudgetOverviewRequest` in the `Common.Models.Budget` namespace, which represent different aspects of a budget in the application. A new project `BudgetService` was added to the solution file `Finance_Manager_App.sln`, and it was configured with the `net8.0` target framework. Several package references were added to this project, including `Microsoft.EntityFrameworkCore`, `Microsoft.EntityFrameworkCore.SqlServer`, and `Swashbuckle.AspNetCore`.

The `BudgetService` project also saw the addition of a new `BudgetController` class, a `DataContext` class, a `DbService` class, and an `IDbService` interface, all of which are crucial for handling HTTP requests related to budgets and interacting with the database. The `Program.cs` file was updated to configure the HTTP request pipeline and add services to the container.

The application's launch settings and logging settings were updated in the `launchSettings.json`, `appsettings.Development.json`, and `appsettings.json` files. The `CategoryController` class was renamed to `BudgetController` in the `UsersService.Controllers` namespace. Lastly, the `DbService` class in the `CurrencyService.Db` namespace was updated to improve code formatting and readability.

List of changes:

1. Addition of `Budget`, `BudgetOverview`, and `BudgetOverviewRequest` classes in `Common.Models.Budget` namespace.
2. Changes to the project `ExpensesSerivce` and addition of `BudgetService` in `Finance_Manager_App.sln`.
3. Configuration of `BudgetService` project with `net8.0` target framework and addition of several package references.
4. Addition of `BudgetController` class in `BudgetService.Controllers` namespace.
5. Addition of `DataContext` and `DbService` classes, and `IDbService` interface in `BudgetService.Db` namespace.
6. Updates to `Program.cs` file.
7. Updates to `launchSettings.json`, `appsettings.Development.json`, and `appsettings.json` files.
8. Renaming of `CategoryController` class to `BudgetController` in `UsersService.Controllers` namespace.
9. Updates to `DbService` class in `CurrencyService.Db` namespace for improved readability.